Removing duplicated limits file and refactored references
