+++ pkg/querier/testutils.go
@@ -3,7 +3,7 @@ package querier
 import (
 	"github.com/grafana/dskit/flagext"
 
-	"github.com/grafana/loki/pkg/util/validation"
+	"github.com/grafana/loki/pkg/validation"
 )
 
 func DefaultLimitsConfig() validation.Limits {
